requirements.txt
flask
google-generativeai
requests
pydantic
